Time: 16 ms (8.51%) | Memory: 8.5 MB (85.64%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     using i64 = long long;
     int mySqrt(int x) {
-        for (i64 i = 0; i < INT_MAX; i++)
-            if (i * i <= x && (i + 1) * (i + 1) > x)
+        for (int i = 0; i < INT_MAX; i++)
+            if ((i64) i * i <= x && (i64) (i + 1) * (i + 1) > x)
                 return i;
         return -1;
     }
